requirements.txt
Flask==2.0.2
numpy==1.26.4
scikit-learn==1.2.2
Werkzeug==2.0.2
pymongo==4.7.2
python-dotenv==1.0.1
flask-cors==4.0.1
snowflake-connector-python==1.5.1